+++ examples/native/scenarios/subgraph_live_query.rs
@@ -2,7 +2,7 @@ use std::{env, error::Error, io};
 
 use serde_json::json;
 
-use cow_sdk::SupportedChainId;
+use cow_sdk::prelude::SupportedChainId;
 use cow_sdk_subgraph::SubgraphApi;
 
 #[tokio::main]